Some new logic and some changes (will work on it)

+++ ColonySimulator.Backend/Migrations/ColonySimulatorContextModelSnapshot.cs
@@ -40,8 +40,8 @@ namespace ColonySimulator.Backend.Migrations
                     b.Property<double>("RequiredStrength")
                         .HasColumnType("REAL");
 
-                    b.Property<double>("ResourceConsumption")
-                        .HasColumnType("REAL");
+                    b.Property<int>("ResourceConsumption")
+                        .HasColumnType("INTEGER");
 
                     b.Property<int>("Strength")
                         .HasColumnType("INTEGER");
@@ -78,8 +78,8 @@ namespace ColonySimulator.Backend.Migrations
                     b.Property<double>("RequiredStrength")
                         .HasColumnType("REAL");
 
-                    b.Property<double>("ResourceConsumption")
-                        .HasColumnType("REAL");
+                    b.Property<int>("ResourceConsumption")
+                        .HasColumnType("INTEGER");
 
                     b.Property<int>("Strength")
                         .HasColumnType("INTEGER");
@@ -116,8 +116,8 @@ namespace ColonySimulator.Backend.Migrations
                     b.Property<double>("RequiredStrength")
                         .HasColumnType("REAL");
 
-                    b.Property<double>("ResourceConsumption")
-                        .HasColumnType("REAL");
+                    b.Property<int>("ResourceConsumption")
+                        .HasColumnType("INTEGER");
 
                     b.Property<int>("Strength")
                         .HasColumnType("INTEGER");
@@ -154,8 +154,8 @@ namespace ColonySimulator.Backend.Migrations
                     b.Property<double>("RequiredStrength")
                         .HasColumnType("REAL");
 
-                    b.Property<double>("ResourceConsumption")
-                        .HasColumnType("REAL");
+                    b.Property<int>("ResourceConsumption")
+                        .HasColumnType("INTEGER");
 
                     b.Property<int>("Strength")
                         .HasColumnType("INTEGER");
@@ -189,8 +189,8 @@ namespace ColonySimulator.Backend.Migrations
                     b.Property<double>("RequiredStrength")
                         .HasColumnType("REAL");
 
-                    b.Property<double>("ResourceConsumption")
-                        .HasColumnType("REAL");
+                    b.Property<int>("ResourceConsumption")
+                        .HasColumnType("INTEGER");
 
                     b.Property<int>("Strength")
                         .HasColumnType("INTEGER");
@@ -227,8 +227,8 @@ namespace ColonySimulator.Backend.Migrations
                     b.Property<double>("RequiredStrength")
                         .HasColumnType("REAL");
 
-                    b.Property<double>("ResourceConsumption")
-                        .HasColumnType("REAL");
+                    b.Property<int>("ResourceConsumption")
+                        .HasColumnType("INTEGER");
 
                     b.Property<int>("Strength")
                         .HasColumnType("INTEGER");
